Refactor: cambiar namespace y corregir importaciones
