+++ app/Http/Resources/Teams/TeamLocalResource.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Resources\Team;
+namespace App\Http\Resources\Teams;
 
 use Illuminate\Http\Resources\Json\JsonResource;
 use App\Http\Resources\Events\EventResource;
