+++ app/Http/Resources/Teams/TeamResource.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Http\Resources\Team;
+namespace App\Http\Resources\Teams;
 
 use App\Http\Resources\Common\LeagueResource;
-use App\Http\Resources\Team\ManagerResource;
+use App\Http\Resources\Teams\ManagerResource;
 use App\Http\Resources\Players\PlayerTeamResource;
 use App\Http\Resources\Whereabouts\CountryResource;
 use Illuminate\Http\Resources\Json\JsonResource;
