+++ app/Http/Resources/Teams/TeamVisitorResource.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace App\Http\Resources\Team;
+namespace App\Http\Resources\Teams;
 
-use App\Http\Resources\Whereabouts\CountryResource;
 use Illuminate\Http\Resources\Json\JsonResource;
+use App\Http\Resources\Events\EventResource;
+use App\Http\Resources\Teams\TeamResource;
 
-class ManagerResource extends JsonResource
+class TeamVisitorResource extends JsonResource
 {
     /**
      * Transform the resource into an array.
@@ -17,10 +18,8 @@ class ManagerResource extends JsonResource
     {
         return [
             'id' => $this->id,
-            'name' => $this->name . ' ' . $this->surname,
-            'birthdate' => $this->birth_date,
-            'picture' => $this->picture,
-            'country' => new CountryResource($this->country),
+            'event' => new EventResource($this->event),
+            'team' => new TeamResource($this->team),
         ];
     }
 }
